(PUP-5209) Fix so that dash can be used in module dependency
The module loader would not find a dependent module when a dash was
used instead of a slash to separate the author and module name in a
dependency in the metadata.json file. Both constructs are allowed
in other places in the code (and on the forge). This commit makes
the loader work with both.