(PUP-5640)(PUP-5591) Fix Win32 Registry Reads which cause errors
Without this patch applied, reading REG_BINARY or corrupted registry
values would cause fatal errors to be raised and the error messages
did not contain enough information for users to remediate.

This commit fixes a typo in the REG_BINARY logic which was referencing
a now non-existent local variable.  Also, the read function
now has a wrapping begin-rescue clause to catch buffer overrun error
messages caused by corrupted registry values e.g. a DWORD having less
than 4 bytes of data.  When a corrupted registry key is found, a warning
is raised stating the key name, not value name, where the value is
located and returns nil for the data of the value.

The tests for this class have been moved from unit to integration as
they were not true unit tests and required actual read/write operations
on the Windows registry.  Tests for the REG_BINARY and corrupted registry
values scenarios were added.