(PUP-5612) Detect and report resource definitions with missing title
A resource declared without title becomes a BlockExpression with
two statments, a QualifiedName and a LiteralHash. This commit adds
validation logic that detects and reports such statements as a new
`RESOURCE_WITHOUT_TITLE` issue. It also ensures that when such a
declaration is found last in a node, define, or host class, it's
reported instead of the `IDEM_NOT_ALLOWED_LAST` since reporting both
would be confusing.