(PUP-5209) Ensure correct visibility between modules
This commit deals with function visibility between modules. It fixes
two major problems and adds a test matrix containing 27 distinct
test cases.

First issue was that a puppet function didn't use the private loader.
This prevented puppet functions in general from calling functions in
other modules.

Second issue was that calls made from outside of a function (i.e.
from .pp code not associated with a 'pops' loader) had access to
all functions in all modules. This happened because the caller
defaulted to using the private environment loader. This commit adds
logic that will attempt to find out what module a call originates from
and, if successful, use the private loader of that module to satisfy
the function lookup. A better fix for this would be to use the new
loader framework throughout but that's a much larger task (see
PUP-1833).

The test matrix tests calls from three different call locations:
- directly from manifests/init.pp
- from a puppet function declared under the modules functions directory
- from a puppet function declared in the modules manifests/init.pp
- from a ruby function

Each call location performs a call to the dependent module that targets
- a puppet function declared under the functions diretory
- a puppet function declared in manifests/init.pp
- a ruby function

Each call is tested using the following conditions:
- Metadata present with a valid dependency to the callee
- Metadata present that lacks a dependency to the callee (should fail)
- No metadata present (will succeed since all modules are visible)