(PUP-5027) Prevent that metadata.json is parsed multiple times
This commit ensures that the number of system calls needed to determine that
a metadata.json file exists and to load it, is kept to a minimum. The check
to see if the file exists was removed. Instead, an attempt is made to read
the file directly and the ENOENT error that might be the result is rescued.

The file is also read and parsed only once.