(PUP-5617) fix test on windows
The original test casted the Socket::Option to a bool and tested the
output. But this raised a TypeError on windows. Additionnally,
Socket.getsockopt() and Socket::Option.new() don't return the same data
on windows.

This patch casts it to a string instead, compares against a real
Socket::Option object, and takes care of the Socket.getsockopt()
platform discrepency.