(PUP-5461) Prevent that lookup evaluates site manifest
Prior to this commit, the `lookup` application would compile a full
catalog which gave undesired side effects. This commit ensures that
this compilation does not take place unless the flag `--compile` is
given on the command line. Without this option, the lookup command will
instead compile the string `undef` to produce the scope needed
for the lookup.