(PUP-5584) Force ASCII-8BIT encoding for cached catalogs
Previously, puppet would attempt to treat JSON read from a cached
catalog as part of the system encoding. This could cause oddities with
how inlined file content is treated.

With this change, catalogs on disk will be treated as ASCII-8BIT. This
essentially tells Ruby to just treat it as a bunch of bytes,
preserving file content and avoiding any encoding issues.