(maint) Allow Function in Block contained by a Program
A parsed file becomes a Program with a Block which means that things
contained in the file resides in a block. Apparently this is not
the case when parsing a string (no block present in the program).

This means that the rule stipulating that a function cannot reside in
a block had to be altered. It's OK for a function to reside in a block
when the container is a program.