(PUP-5658) Disallow numeric types with ranges where to < from
Numeric types like `Integer[10,0]` was allowed. This commit ensures
that an error is raised if 'to' is less than 'from' in the range.

**Important note when merging to 4.x**
The checks for to < from (here in multiple places) should NOT be
merged. In 4.x there should be only one check and it should reside
in the `PNumericType.initialize`.