(PUP-1985) Fix correct scope for parameter default evaluation
This commit ensures that a correct scope is used when evaluating
parameter default expressions. It impacts the parameter declarations
of functions, EPP templates, classes, and resources.

A new ephemeral scope `ParameterScope` was added to keep track of
which parameters that exists and in which order they are evaluated
and accessed. This scope is only in effect during the evaluation
of the parameter defaults. Since the parameters are evaluated left
to right, and each parameter is added to the scope, a default
expression of a parameter can use the parameters declared to the
left.

Each parameter in turn is evaluated using a private MatchScope to
ensure that the evaluation of one parameter cannot interfere with
another. No LocalScope is pushed though, since it's forbidden
to assign new variables in this context.

The commit also contain three new test suites (for function, epp-
template, and class+resource parameters). They all test default
parameter evaluation much in the same way but for different targets.