(PUP-1607) Acceptance - no master in symbolic_modes
This commit confines hosts eligible to be included in the execution
of the `resource/file/symbolic_modes` test.

It removes the master from the hosts list.
It also moves the logic to exclude windows and solaris-10 from
the body of the test and places them as confine statements at
the beginning.