(PUP-5583) Use beaker to install puppet-agent
Previously, `install_repos_on` was called for all host platforms,
which is a helper in puppet, and needed updates for each new agent
platform.

This commit bumps our Beaker dependency to use the updated
`install_puppet_agent_dev_repo_on` method, and uses that method to
install puppet-agent on all agent-only hosts.

For the master, we install the puppet-agent and puppetserver repo
configs, and install puppetserver, which installs puppet-agent as a
dependency. It is important we don't install puppet-agent on the
master host before puppetserver due to SERVER-528.

Unfortunately, `install_repos_on` and related methods cannot be
removed from puppet because beaker cannot install the latest
puppetserver build from nightlies due to BKR-673.