(PUP-5106) Add catalog_uuid identifier to catalog
Currently, every agent run generates a unique transaction_uuid which
is attached to the catalog and report after the run has finished.
This allows the catalog to be linked to a specific report. However,
in the case of running off of cached catalogs, a single catalog can
have multiple reports.

When this happens, the catalog retains its original transaction_uuid
while the new reports contain the newly generated UUID, making it
impossible to link the cached catalog to any reports except for the
very first one.

This commit adds a new `catalog_uuid` identifier to the catalog,
which will allow a single catalog to be linked to multiple reports.