(PUP-5713) Let ClassInformationService use source of parameter defaults
A parameter default that is not a literal must be represented as
source in the info structure created by the ClassInformationService.
A bug in the SourcePosAdapter where all contained elements were
assumed to have a non-nil offset prevented this function.

This commit fixes the bug and adds a test to verify that the the
source is extracted correctly.

The commit also changes the behavior so that the 'default_source'
entry will always be present in the info structure when a default
expression is present, regardless of if it also contains a
'default_literal' entry.