(PUP-5694) Allow untruthy environment values to override truthy ones
Previously, if Puppet[:static_catalogs] was true, but environment.conf
specified a false value, then the resulting value was always true.  This
was because we required the overridding value to be truthy, when you may
not want that. For example, to set static_catalogs to true for all
environments, except production.

This commit makes it so we only fall back to the default case if the
value is nil.