(PUP-5729) Add iterative reverse_each function
This commit adds the new function `reverse_each(iterable)`. The
function takes an iterable argument and an optional block.

If a block is given, it will be called with all elements of the given
argument in reverse order.

Without a block, the function will return a new Iterable that will
yield each element of the given argument in reverse order.