(PUP-5730) Add iterative step function
This commit adds the new function `step(iterable, n)`. The
function takes two arguments, an iterable and an integer, and an
optional block.

If a block is given, it will be called with each nth element of the
given argument.

Without a block, the function will return a new Iterable that will
yield each nth element of the given argument.