(PUP-917) Identify in the report why a cached catalog was used
Currently, the agent run report does not provide any information
about why a cached catalog was used. This can cause issues when
a remote catalog request unexpectedly fails, causing the agent
to fall back to a cached catalog. In such a case, the report will
indicate that the run was successful without pointing out that a
cached catalog had to be used.

There are two ways a cached catalog may be used: by running the agent
with the `--use_cached_catalog` option, or by a catalog failing to
compile on the master, causing the agent to automatically fall back to
a cached catalog as long as the `usecacheonfailing` setting is set.

This commit adds the `cached_catalog_status` field to Puppet::Transaction::Report,
with possible values 'unused', 'use_cached_catalog' and
'use_cache_on_failure'.