(PUP-5120) Send checksum_type and static_catalog in catalog request
Currently, when the agent applies a catalog, it sends a `file_metadata`
request for each file resource along with the `checksum_type` that the
master should use to compute the checksum. The `checksum_type` originates
in either the manifest (and preserved in the catalog), or if none was
specified, the agent's default `Puppet[:digest_algorithm]`.

When static catalogs are enabled, the file metadata inlining process
will need to choose a digest algorithm if none is specified in the
manifest. To ensure the master chooses a checksum suitable for the
agent, the agent needs to send its desired checksum in the catalog
request as a request parameter. Add a checksum_type query param to the
catalog request with with the value of `Puppet[:digest_algorithm]` to
select the preferred checksum_type for file resources.

Also, request a static catalog by sending `static_catalog=true` query
parameter. The server will decide whether to build a static catalog.