(maint) Remove Ruby 2.3.0 from Travis C.I. matrix
Since we're getting a mysterious seg fault when running our spec
tests against Ruby 2.3.0, remove it from the Travis testing matrix
for now until the issue is sorted out.

The issue in question is: https://bugs.ruby-lang.org/issues/11928
which has been fixed. Once the new version of Ruby is released
with this fix, we should be able to add 2.3.0 back to the testing
matrix.