(PUP-5708) Deprecate pluginsync setting
Soon, the behavior of pluginsyncing will be based on whether or
not a cached catalog is being used. This is to prevent scenarios
where the set of plugins used by the agent is different than the
set that was used when the catalog was originally compiled.

This commit deprecates the setting. Future work will handle the
switchover to the new logic based on cached catalog usage.