(PUP-5718) Update puppet.bat to current standard
With the switch over to building puppet-agent for windows with Vanagon,
we need to effectively install puppet with install.rb. This script
installs puppet.bat, which is a wrapper that lets windows handle ruby
scripts. This commit updates puppet.bat to use our environment.bat file if
we install it, and to be more in line with the puppet.bat file installed
with the puppet-agent package as built by puppet_for_the_win