(PUP-5745) Use PuppetDB 4 API for querying where applicable
Previously, the CapabilityFinder was using the
Puppetdb::Util::Http.action(...) method to query PuppetDB. In PuppetDB 4
however, the action() method has changed to require a new optional
argument, which breaks compatibility. At the same time, a top-level
query_puppetdb() method has been introduced which is intended to be the
public API going forward. We now check for and use the query_puppetdb()
method where it exists, falling back to the old internal action() method
when necessary.