(PUP-5648) Add Iterable and Iterator to Puppet Type System
This commit adds the new types `Iterable` and `Iterator` to the Puppet
type system together with the runtime support needed to represent an
object that has been transformed into an Iterator by a call to
`Iterable.on`. This method transforms an instance of an object that
Puppet can use in iterative functions into an Iterator. The Iterator
in turn, provides basic support for the iterative functions.