(PUP-5741) Fix confusing errors when comparing values
The method MessageData#label has two functions. If called with an
argument, it presents the String form for that argument. If called
without, it returned a `ModelLabelProvider`. Unfortunately, it decided
whether or not it got an argument by testing if it's value was `nil`
(the default). This made it impossible to produce a message for the
argument `nil`.

This commit changes the parameter so that it uses the splat operator
and can check the actual number of passed arguments.