(PUP-3077) Don't cache catalogs during noop runs
Previously, agents always cached their catalogs in
$client_datadir/catalog/<host>.json, even when running in noop mode. If
the agent failed to download a catalog during the next run, e.g. puppet
server was stopped or overloaded, etc, the agent would fall back to the
cache, due to the usecacheonfailure setting, and apply the previously
cached catalog. This could lead to surprising results if you were just
testing out the noop run and never meant for those changes to be
deployed in production.

This commit will set the catalog's cache to nil in noop mode so that the
catalog is never cached. It also changes the apply application to behave
the same.

The agent_spec explicitly unstubs `cached_class=`, because mocha behaves
unintuitively when setting a `never` expectation on a previously stubbed
method[1]. This was noticed, because the test wasn't failing as expected
when run against the old code.

[1] https://github.com/freerange/mocha/issues/44#issuecomment-3090212