(maint) Don't stub private method implementations
The tests were stubbing the implementation of `convert_catalog` so that
it always returned the JSON catalog passed to the method. However, in
reality, the method returns a RAL catalog. As a result, the tests were
testing non-existent behavior.

This commit moves the convert_catalog call outside retrieve_catalog, so
that we test actual behavior.