(PUP-3077) Ignore cached catalog if environment doesn't match
Previously, if puppet failed to download a new catalog and fell back to
its cache (due to usecacheonfailure), it would load and apply the cached
catalog, even if the catalog's environment didn't match its configured
environment (as specified by the ENC).

With this commit, if the agent fails to download a new catalog, and
falls back to its cache, and the cached catalog's environment doesn't
match the agent's current environment as specified by the server, then
the agent will print an error and fail the run.

This commit does not change the behavior when the agent is told to use
its cached catalog to start with (via use_cached_catalog) or if the
server doesn't specify an environment to use. In that case, the
agent will load the cached catalog, and switch to the catalog's
environment if need be.

Puppet will now include the environment name in the 'Using cached catalog'
info message.

Although the change is in configurer, it should only affect the agent
and not apply, because the latter always compiles a new catalog.

Some tests were updated to ensure catalogs were created with a valid
environment instead of defaulting to the NONE environment.