(PUP-5538) Improve SID::Principal.domain_account validation
Previously the lookup_account_sid and lookup_account_name functions
would just append the account name to the domain name if a domain name
was returned.  However this caused bad domain_account attributes such as
'BUILTIN\BUILTIN' when querying the BUILTIN domain.  This commit changes
the Principal domain_account attribute initialization to only use the
domain name for SidTypeDomain principals, if not the previous behavior of
domain and account names is used.  This commit also adds tests for
querying the BUILTIN domain.