(PUP-5744) Add a JsonStrictLiteralEvaluator
Before this, there was only a LiteralEvaluator and it considers all
Puppet literals as literals. This is a problem in cases where the intent
is to present the value as "pure json" (i.e. wher hashes only accepts
string keys, and where regular expressions and the concept of default
cannot be encoded).

This adds the class JsonStrictLiteralEvaluator which only produces
literal values that can be represtend in pure json.