(PUP-5744) Make class_information_service produce json-safe literals
Before this, literal default values returned for parameter default
expressios included values that cannot be represented as pure json. The
unsafe literals Regexp, instances of Default type, and instances that
are not of Hash[String, Any] type did all cause problems when
transformed to json.

Now, the json safe literal value evaluator is used, and usafe json
literals are not included in the response. For unsafe values, the source
text must be used instead.