(PUP-5802) Make the Yum package provider correctly parse epochs
Without this patch applied, the Yum package provider assumes all version
epochs are "0", i.e. a package with version "1:1.2.3-4.el5" will be
interpreted as having epoch "0" not "1". This causes version comparisons
made by puppet to be incorrect when comparing packages of different
epoch's.

This bug is caused by the line:
 e = String(Bignum(e)) rescue '0'

Bignum is not a function, so the line always throws an exception and the
rescue is always invoked.

This patch uses Integer instead of Bignum, rescues the specific
exception needed (this is so that an epoch like "foo" is coerced into
"0"), and provides a test for non-zero epochs