(PUP-5699) Inline symlink metadata in static catalog
This commit updates the static compilation inlining logic to
handle symlinks which have their `links` parameter set to manage.
The `source` parameter of such a resource references one of two
items: a link or a regular file.

When `source` references a link, the destination of that link is
inlined directly into the `target` parameter of the new resource.

If `source` references a regular file, the existing logic holds
and file content is copied into the new resource.

When `links` is set to `follow`, we fall back to the regular
content inlining behavior, which sets the `checksum_value` of
the new resource.