(maint) Return a catalog in the current environment
We were returning a catalog with the default "none" environment, which
didn't match the current environment. The configurer would retry the new
catalog catalog request to converge, which is why the test was expecting
twice.

The behavior is already tested in "should refetch the catalog if the
server specifies a new environment in the catalog", so no need to test
it again.