(PUP-1780) Allow runtime3 to optionally fail on undefined variable
Although not needed for PUP-1780, this handles the case when undefined
variable is always thrown, and the --strict_variables defaults to
true. (Then, all attempts to get an udefined variable will result in a
throw of :undefined_variable)

In the fix for PUP-1780, this approach was tried first, and the test
errors that followed due to the changed API revealed bugs (fixed in
subsequent commits). That solution could however not be used as it
requries a major version release

In this commit, the handling of the UNKNOWN_VARIABLE issue is configured
to warning or error depending on the --strict_variables flag.