(PUP-1780) Add a warn_once to Logging
Before this, it was not possible to issue a warning only once without
using a deprecation_warning. That is a problem because deprecation
warnings can be excluded, and they are capped at 100 warnings. This mas
that to fix undefined variable problems you would also have to fix all
other deprecations - or you may not even see the undefined variable
warnings!

It would be ok to use puppet_deprecation_warning otherwise, but from a
pure naming standpoint, the warnings does not have to be deprecations as
the puppet_deprecation_warning does not add anything to the message
(like "deprecation"). It does however make the code read in a bad way 

The new `warn_once` accepts a `kind` which can be used in puppet's
support for exclusion of warnings. This means that users can exclude
other deprecations while not excluding `unknown_variable`.

In this commit is also a small utility that makes testing against
multiple variables easier. Ideally this could be handled by the puppet
type system, but that is not suitable for Logging as it needs to be able
to log early in the lifecycle of the Puppet runtime. (A ticket was added
to support puppet type system types in case expressions)