(PUP-1780) Make Scope issue warning for all non existing variables
Before this, only variable lookup of multi segment qualified variables
resulted in a warning for a non existing variable.
This is bad because it makes it very hard to find and correct all places
where undefined variables are used in preparation of turning on --strict
variables.

Now, Scope issues a warning for every lookup where the variable does not
exist.

It is possible to use the :disable_warnings setting to disable all
'undefied_variables' warnings.