(PUP-5819) Make lexer error when seeing a Byte Order Mark first in input
Before this, a byte order mark at the begining of the input would cause
the lexer to deliver the unsupported characters as an OTHER token. This
leads to a Syntax Error.
This is bad because some of the BOMs are non printable characters and
users will not know what is wrong. Especially when files having been
edited on a Windows platform, and later cause problem on a n*x box where
users are less familiar with seeing BOMs.

This PR adds an assertion when the lexer is about to deliver an OTHER
token. This BOM assertio checks if the starting 2-4 bytes of the input
is one of the 10 BOMs in wide use. When it errors it will show the name
of the format e.g. UTF-16, or UTF-32, and the bytes that defines that
BOM. The user is enciuraged to remove this sequence at the beginning of
the file in the error message.

Ruby, after first not supportig wide characters shifted and instead made
it difficult to work with the raw bytes. Some gymnastics are required to
represent and process the characters as several of the sequeces form
invalid strings in the UTF-8 encoding we are using in puppet source.