(PUP-5701) Refactor file type so source implements writing
Previously writing data for a file to disk required that a content
property be defined. Future modifications to the file type will use
checksum_value to determine whether an update needs to occur, and may
prefer content is not defined.

Moves implementation of writing data from source to the source
parameter, rather than delegating through a content property (which
is no longer assumed to exist). Content still triggers the write.