(PUP-5701) Use checksum_value instead of content if specified
If the checksum_value property is specified, use it instead of content
for verifying whether a file is insync. Changes checksum_value to a
property, and triggers file writing if both checksum_value and source
are specified; in that case it will prevent a content property from
being created from source.