(PUP-5824) Support empty interpolation in Hiera data provider
Before this, an empty interpolation such as `%{}` was not considered
an interpolation at all. The interpolation regexp required at least one
character between the curly braces.

This commit will allow the empty interpolation and will also ensure
that the expression between the curly braces is trimmed from whitespace
at both ends. If the result of that trim yields the empty string or
a string that only contains a name separator '::', then the interpolated
result will be the empty string.

It is still possible to have interpolated expressions where the key has
leading or trailing whitespace by using the function syntax, e.g.
%{hiera(' key ')}.