(PUP-5835) Make node params String if they are Symbol
Before this, if a node parameter had a Ruby Symbol value it would leak
to Scope, and into variables. This was especially true for the
$environment variable after an enviroment name is made into a Symbol.

This commit fixes the problem both by:
* ensuring that when a node sets a
parameter (in itself) based on the environment's name it transformes the
symbolic name to a String.
* ensuring in the general that the compiler, when
setting parameters for a node, will also transform a Symbol value to
String. The later to ensure the same kind of mistake does not leak out
to Scope in the future.

This commit also includes unit tests that asserts that environmet name
as a Symbol does not leak, nor any other node parameter value.