(PUP-5836) Make splat create Array from Iterator
Before this a splat on an Iteraor had no effect - it resultet in the
Iterator (as it was an atomic value).

Now the splat operator will produce an intance of Array from the
iterable. If used in a context where the splat means that the array is
ufolded, it will also be unfolded.