(PUP-5742) Add Recursion guard to the Puppet Type system
This commit adds a RecursionGuard class and changes the signature of
methods in the Puppet Type System that will be subjected to type aliases
that in turn might contain self recursion.

The guard is designed to keep track of two chains where a possible
self recursion can occur called 'this' and 'that'.

The guard is not really used in this commit since none of the current
types can contain self recursion and excessive checks for self recursion
in the type system might have very negative performance implications.