(PUP-5742) Add TypeAlias to the Puppet Type system
This commit adds the PTypeAlias class and associated factory and print
methods. It also makes the TypeParser aware of the existence of type
aliases. The parser will now make an attempt to load a type using the
Puppet::Pops::Loader mechanism before it concludes that the type is
a Resource type. When no scope has been provided (this happens when
the TypeParser is called from the ClassInfoService), then the loader
cannot be used. Consequently, it cannot be determined if an unknown
name represents a Type alias or a Resource type and the parser will
now return a TypeReference instance when that happens.