(PUP-5742) Move and rename method PopsBridge::module_loader
This commit moves the method PopsBridge::module_loader to the
Puppet::Pops::Loaders class and renames it to `find_loader` (since
it's applicable without a module name too). It also divides the method
in two. One on the Loaders class, and then, when a Loaders instance
is found, it delegates to that instance.