(PUP-5119) Only inline metadata if the source is in the environment
This adds a check around replacing the resource's metadata to verify
the file is versionable (exists within the environment path tied to
the catalog's code id). Doing this at the compiler layer greatly
simplifies the code compared to trying to forward all the needed
parameters to the file server layer. This does mean we occasionally
stat a file we won't inline, but that's worth it (until we see
real-world numbers saying it's expensive) to avoid making this super
complicated.