(PUP-5698) In-line directory metadata
Update the catalog compiler so it inlines directory metadata for
static catalogs. This means that when there is a file resource
that is a directory and has recurse set to true, we will add resources
for all it's children to the catalog.

In order to support this, make a few changes to the file type as
well so that we can easily reuse functionality for static catalogs.