(PUP-5861) Make Array produce 'Array' when presented as a String
Before this, an Array created without type parameters would present
itself in string form with the Data element type and 0 (or more length).
This is surprising to users as that is not what Array looks like in
code.

This changes the string representation of the Data compatible Array to
be just Array.