(PUP-5861) Make Hash produce 'Hash' when presented as a String
Before this, a Hash created without type parameters would present
itself in string form with the Scalar key, and Data value type and 0 (or
more length).
This is surprising to users as that is not what Hash looks like in
code.

This changes the string representation of the Data compatible Hash to
be just Hash.