(PUP-5805) Set missing location in generated hash
Before this an expression like:
```
notice {a=>2}
```
Would result in an internal error.

That was caused by the "tree rewrite" that changes that expression from
being a resource expression without title to a function call with a hash
as an argument. When the hash had no location information it lacked
offset and length of its textual representation in source text.

This is now fied by propagating the tokens for the left and right braces
to the rewrite operation that now sets them.