(maint) Use compile_to_manifest and resource matchers
Rewrite tests to use `compile_to_manifest` so it's clearer what we're
compiling. Also use resource matchers to assert state about inlined
resources.

Adds a test for a directory when recurse not specified, and tests for
non-file resources and before/require relationships which were
previously being tested indirectly via `build_catalog`.

Adds a pending test for a director when recurse is true, as ensure is
not set on the children nodes.