(maint) Don't return the catalog from inlining
The inline_metadata method modifies the catalog it is given, so we don't
need to return it back to the caller. It avoids confusion about whether
the returned catalog is a different instance than what was passed in. It
also removes asserting that the returned catalog was equal since that
was always true.