(maint) Add pending test for windows paths
Calling `Resource#to_ral` on a *nix OS for a windows path will fail,
because 'c:/foo' is not an absolute *nix path. This commit adds a pending
test for this test. It shows that calling `to_ral` during compilation is
a bad idea, because agent-specific validation can occur on the master.
Fortunately, the file type doesn't call into the provider during
validation or try to load platform specific code.