(PUP-5837) Change reverse_each to return undef when given a block
Before this commit, the `reverse_each` function would return its first
argument when given a block. This commit changes that so that it returns
the undef value instead to avoid any confusion regarding if the returned
value is reversed or if it somehow contains values returned from the
given block.