(PUP-4818) Make class references absolute for resource like include
Before this, when using future parser, where dynamic namespacing is
supposed to have been removed (as in 4.x), and using a resoure like
include expression for a class the search for the class takes the name
as a relative name unless i starts with ::

```
class { 'foo': }
```
finds the relative foo, not ::foo.

This commit changes this so that all references are taken to be absolute
if they are not already.