(PUP-5118) Refactor remove_less_specific_files
Refactor the remove_less_specific_files method so that it can be
used not only by the file type but also by the static compiler to
handle recursive file resources when in-lining metadata. It's used
to ensure we do not have a duplicate resource for a file in the
case where it is managed explicitly and also covered by its parent
directory which has recurse turned on.