(PUP-4818) Correct grammar error where token instead of value is used
Before this, the grammar created a QNAME for a token, instead of the
tokens value. This led to an attempt to call a function named "'type
TYPE'".

The grammar should have used val[0][:value] to get the token text, but
used only value[0], which is the token itself.

This is now corrected.