(PUP-5735) Remove usage on Win32-Security Gem
The functionality of the Win32-Security gem has been merged into Puppet.
This commit moves the last of enumerations and updates the test suites to
use the newer namespace of Puppet::Util::Windows:SID.  This commit removes
Puppet's requirement of the win32-security gem, however the gem itself
still exists in the ruby gem library.  It will be removed from the library
in later commits.