(PUP-5697) Inline file permissions to the static catalog
Inlines `owner`, `group`, and `mode` permissions when
`source_permissions` is `use`. Not inlined when it's `ignore`, and
raises an error if `use_when_creating` is requested.

For Windows, modified how we assert that `source_permissions => use`
will not be used on Windows so it's still checked for the static
catalog. Check is done client-side, as the master usually doesn't care
what OS the client is.