(PUP-5761) Fix so that hiera module data provider validates correctly
Prior to this commit, the hiera module data provider did not validate
its data. The reason for that was that the underlying file based data
providers had no knowledge that they were sub-providers to a hiera
module data provider.

This commit changes the PathBasedDataProvider (super class of the
file based one) so that delegates the data validation to its parent
provider.