(PUP-5761) Make PathBasedDataProvider aware of data provider container
Before this commit, the PathBasedDataProvider was not aware of its
parent data provider. This makes it impossible for the path based
provider to validate its data since it doesn't know if that data
is module data or something else.

This commit prepares for such validation by making the data provider
"parent aware". In order to do this, an extra argument must be passed
to the bound factory's create method which of course breaks API. This
was resolved by introducing a 'version' attribute on the factory and
having the caller first test if the factory responds to version. If it
doesn't, it's considered to be version 1 with the old API. If it does,
then the create method will accept a third parameter (the parent data
provider).

The HieraConfig.create_factories method that takes one argument
was deprecated in favor of a new create_configured_factories method
where the parent data provider is added as a second argument.

The change adds to the API but should be fully backwards compatibly.